iio:xadc: Add workaround for PS7 event handling
The PS7 XADC threshold interrupts are level sensitive. Since we can't make the
threshold condition go way from within the interrupt handler, this means as soon
as a threshold condition is present we would enter the interrupt handler again
and again. To work around this we mask all active thresholds interrupts in the
interrupt handler and start a timer. In this timer we poll the interrupt status
and only if the interrupt is inactive we unmask it again.

[Lars-Peter Clausen: Style cleanups, minor fixes, use delayed work]
Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>